+++ src/@types/_index.ts
@@ -1,3 +1,4 @@
 export * from "./commonView";
 export * from "./view";
 export * from "./positionTypes";
+export * from "./errorBoundary";